+++ src/index.js
@@ -1,1 +1,1 @@
-export { default as LoginForm, makeTheme as makeThemeLoginForm, theme as themeLoginForm } from './LoginForm';
+export { LoginForm, makeTheme as makeThemeLoginForm, theme as themeLoginForm } from './LoginForm';